Voice activated agent
This builds off the "TPMA_Workshop_agent" to connect Telegram and allow you to send text or voice messages to your agent to receive responses. This is a great step if you are looking for how to add simple voice activation for your agent. It leverages the Open AI transcription node to transcribe the data you sent with a voice message in Telegram.